+++ board/recalbox/fsoverlay/recalbox/scripts/recalbox-config.sh
@@ -56,13 +56,6 @@ systemsetting="recalbox_settings"
 recallog -s "${INIT_SCRIPT}" -t "CONFIG"  "---- recalbox-config.sh ----"
 
 if [ "$command" == "getRootPassword" ]; then
-    # security disabled, force the default one without changing boot configuration
-    securityenabled="`$systemsetting  -command load -key system.security.enabled`"
-    if [ "$securityenabled" != "1" ];then
-        echo "recalboxroot"
-        exit 0
-    fi
-
     ENCPASSWD=$(grep -E '^[ \t]*rootshadowpassword[ \t]*=' "${storageFile}" | sed -e s+'^[ \t]*rootshadowpassword[ \t]*='++)
     if test -z "${ENCPASSWD}"
     then
@@ -77,12 +70,6 @@ fi
 
 if [ "$command" == "setRootPassword" ]; then
     PASSWD=${2}
-
-    # security disabled, don't change
-    securityenabled="`$systemsetting  -command load -key system.security.enabled`"
-    if [ "$securityenabled" != "1" ];then
-        exit 0
-    fi
 
     # if no password if provided, generate one
     if test -z "${PASSWD}"
@@ -530,7 +517,7 @@ if [ "$command" == "wifi" ]; then
         do
             # Wait for an IP
             if [ `ifconfig wlan0 | grep -c "inet addr"` -gt 0 ]; then
-                break; 
+                break;
             fi
             sleep 1
             try=`expr $try + 1`