+++ CHANGELOG.md
@@ -121,6 +121,7 @@ All notable changes to this project will be documented in this file.
 
 - rpi3 now uses Mesa software stack (#????)
 - Add required bios for DuckStation (#1944)
+- Removed security option (#1913)
 
 ## [7.2.2-Reloaded]
 - Fix controler #2 in swanstation (#1730)